chat-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
pymongo==4.5.0
pyjwt==2.8.0
sentence-transformers==2.2.2
pinecone-client==2.2.2
torch==2.0.1
transformers==4.33.2